# developer-docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Persona.ai Developer Platform API",
    "version": "1.0.0",
    "description": "REST + AG-UI API for integrating your product with Persona's agent infrastructure.\n\nEvery endpoint here is a server-to-server call — see the Integration Guide for where in your own stack to call each one from, and how the same endpoint serves both Project-level and per-end-user requests via the `x-persona-external-user-id` header."
  },
  "servers": [
    {
      "url": "https://api.persona.hasanraiyan.me",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "projectCredential": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "<keyId>.<secret>",
        "description": "Developer Platform Project credential, minted via POST /api/v1/projects/{projectId}/credentials. Send as `Authorization: Bearer <keyId>.<secret>`. Optionally pair with the `x-persona-external-user-id` header to act on behalf of one of the Project's own external users. This credential is a server-side secret — see the Integration Guide."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/developer/skills": {
      "post": {
        "tags": [
          "Developer"
        ],
        "summary": "Create a Skill (Project-owned or ExternalUser-owned)",
        "description": "Which kind gets created depends on the credential's asserted identity, same as Developer Agent creation.",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "description",
                  "instructions"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Skill created"
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "A Skill with this exact name already exists"
          }
        }
      },
      "get": {
        "tags": [
          "Developer"
        ],
        "summary": "Discover Skills",
        "description": "A genuinely separate code path from Persona's marketplace search. For a bare Project credential (ProjectMachineContext/ ProjectAdminContext): every Skill in this Project's own Domain, any owner type. For a credential paired with x-persona-external-user-id (ProjectRuntimeContext): scope=mine returns only that external user's own Skills; omitting it returns this Domain's public Skills only.\n",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "mine"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of Skills"
          }
        }
      }
    },
    "/api/v1/developer/skills/{skillId}": {
      "get": {
        "tags": [
          "Developer"
        ],
        "summary": "Get a Skill",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "skillId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Skill"
          },
          "404": {
            "description": "Skill not found (or not visible to this Domain/Subject)"
          }
        }
      },
      "patch": {
        "tags": [
          "Developer"
        ],
        "summary": "Update a Skill (owner only)",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "skillId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated Skill"
          },
          "404": {
            "description": "Skill not found or unauthorized"
          },
          "409": {
            "description": "Another Skill with this name already exists"
          }
        }
      },
      "delete": {
        "tags": [
          "Developer"
        ],
        "summary": "Delete a Skill (owner only)",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "skillId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Skill deleted"
          },
          "404": {
            "description": "Skill not found or unauthorized"
          }
        }
      }
    },
    "/api/v1/developer/skills/{skillId}/usage": {
      "get": {
        "tags": [
          "Developer"
        ],
        "summary": "See what's using this Skill, before attempting to delete it",
        "description": "Answers \"what's using this\" proactively, rather than only finding out via a rejected delete call. `agents` is a preview (capped at 20); `agentCount` is the real total.\n",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "skillId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ agentCount, agents: [{ _id, name }] }"
          },
          "404": {
            "description": "Skill not found or unauthorized"
          }
        }
      }
    },
    "/api/v1/developer/skills/bulk-delete": {
      "post": {
        "tags": [
          "Developer"
        ],
        "summary": "Delete multiple Skills in one call",
        "description": "Best-effort batch delete — one blocked/not-found id doesn't abort the rest. Up to 100 ids per request.\n",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ids"
                ],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "maxItems": 100
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "{ deleted: string[], failed: [{ id, reason }] }"
          },
          "400": {
            "description": "ids missing, empty, or over 100 entries"
          }
        }
      }
    },
...
  },
...
}